chore(deps): ⬆️ Updated egen to v1.2.3

+++ frontend/src/rpc/progress.zod.ts
@@ -3,9 +3,9 @@ import { z } from 'zod';
 import { Exchanges } from './exchanges.zod';
 
 export const Progress = z.object({
-  cur: z.number().max(9223372036854775807).min(-9223372036854775808),
+  cur: z.number().int().max(9223372036854775807).min(-9223372036854775808),
   exchange: z.lazy(() => Exchanges),
-  size: z.number().max(9223372036854775807).min(-9223372036854775808),
+  size: z.number().int().max(9223372036854775807).min(-9223372036854775808),
   symbol: z.string(),
 });
