+++ frontend/src/rpc/status.zod.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 export const Status = z.object({
-  code: z.number().max(4294967295).min(0),
+  code: z.number().int().max(4294967295).min(0),
   message: z.string(),
 });
