+++ frontend/src/rpc/symbol-info.zod.ts
@@ -5,12 +5,12 @@ import { SymbolType } from './symbol-type.zod';
 
 export const SymbolInfo = z.object({
   base: z.string(),
-  baseCommissionPrecision: z.number().max(9223372036854775807).min(-9223372036854775808),
-  basePrecision: z.number().max(9223372036854775807).min(-9223372036854775808),
+  baseCommissionPrecision: z.number().int().max(9223372036854775807).min(-9223372036854775808),
+  basePrecision: z.number().int().max(9223372036854775807).min(-9223372036854775808),
   exchange: z.lazy(() => Exchanges),
   quote: z.string(),
-  quoteCommissionPrecision: z.number().max(9223372036854775807).min(-9223372036854775808),
-  quotePrecision: z.number().max(9223372036854775807).min(-9223372036854775808),
+  quoteCommissionPrecision: z.number().int().max(9223372036854775807).min(-9223372036854775808),
+  quotePrecision: z.number().int().max(9223372036854775807).min(-9223372036854775808),
   status: z.string(),
   symbol: z.string(),
   symbolType: z.lazy(() => SymbolType),
